+++ libraries/botbuilder-adapter-twitter/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@botbuildercommunity/adapter-twitter",
-  "version": "1.0.0",
+  "version": "1.0.0-pre.1",
   "description": "Twitter adapter for use with the Bot Framework.",
   "main": "./lib/index.js",
   "typings": "./lib/index.d.ts",
