+++ libraries/botbuilder-adapter-twitter/src/index.ts
@@ -5,4 +5,5 @@
 export * from './schema';
 export * from './twitterAdapter';
 export * from './twitterWebhook';
+export * from './twitterSubscription';
 export * from './util';
