+++ samples/adapter-twitter/now.json
@@ -2,7 +2,7 @@
     "version": 2,
     "name": "adapter-twitter-example",
     "builds": [
-        { "src": "*.js", "use": "@now/node" }
+        { "src": "*.js", "use": "@now/node-server" }
     ],
     "routes": [
         { "src": "^/(.*)", "dest": "/index.js" }
